splitting functionality into different files, cleaning up unused variables and trailing whitespace
