+++ client/src/components/Timeline/index.jsx
@@ -1,119 +1,72 @@
-import { useForm } from "react-hook-form"
+import { useForm } from "react-hook-form";
 
-import { setDefaultOptions } from "date-fns"
-import { fi } from "date-fns/locale"
-import { useEffect, useRef, useState } from "react"
+import { setDefaultOptions } from "date-fns";
+import { fi } from "date-fns/locale";
+import { useState } from "react";
 
-import TimelineFilterForm from "./TimelineFilterForm"
-import { useSchoolQuery } from "@/api/schools"
-import { useParams } from "react-router-dom"
+import TimelineFilterForm from "./TimelineFilterForm";
+import TimelineItem from "./TimelineItem";
+import Timeline from "./Timeline"
+import { useSchoolQuery } from "@/api/schools";
+import { useParams } from "react-router-dom";
 
-import moment from "moment"
-import "moment/dist/locale/fi"
+import moment from "moment";
+import "moment/dist/locale/fi";
 
-moment.locale("fi")
-
-
-const TimelineItem = ({ timeStart, timeEnd, unavailable, user, label, fullHeight }) => {
-  const from = timeStart.split(":").map(v => parseInt(v))
-  const to = timeEnd.split(":").map(v => parseInt(v))
-
-  const columnStart = (from[0] * 4 + Math.floor(from[1] / 15) + 1)
-  const columnEnd = (to[0] * 4 + Math.floor(to[1] / 15) + 1)
-
-  const gridColumnValueString = `${columnStart} / span ${columnEnd - columnStart}`
-
-  if (unavailable) {
-    return (
-      <div
-        className="bg-gray-200 z-10 rounded-md warning pointer-events-none"
-        style={{ gridColumn: gridColumnValueString, gridRow: "1 / -1" }}
-      >
-        {/* <p className="text-ms uppercase font-medium text-black" style={{ position: "sticky", left: "10px" }}>Ei varattavissa</p> */}
-      </div>
-    )
-  }
-
-  const backgroundColor = fullHeight
-    ? "#EA7272"
-    : "#F4BD89";
-  
-  return (
-    <div 
-      onMouseMove={(event) => event.stopPropagation()}
-      className="rounded-md w-full select-none hover:z-10 hover:cursor-pointer overflow-hidden hover:overflow-visible relative"
-      style={{
-        gridColumn: gridColumnValueString,
-        gridRow: fullHeight ? "1 / -1" : "auto",
-        backgroundColor,
-        transition: "all 0.05s linear",
-        transformOrigin: "center center"
-      }}
-    >
-      <div className="bg-inherit py-[2px] px-2 rounded-md absolute min-w-full top-0 left-0 transition-shadow hover:shadow-xl">
-        <p className="font-semibold whitespace-nowrap text-sm bg-inherit max-w-max">{user.name} {user.surname}</p>
-        <p className="text-xs bg-inherit">{timeStart} - {timeEnd}</p>
-        <p className="text-xs max-w-full whitespace-nowrap text-ellipses overflow-hidden bg-inherit"><i>{label}</i></p>
-      </div>
-    </div>
-  )
-}
-
-let trackStart = 0
+moment.locale("fi");
 
 const TimelineContainer = ({
   rooms = [],
   showRoomInformation = true,
 }) => {
 
-  const [ trackingMouse, setTrackingMouse ] = useState(false)
-  const [ trackStart, setTrackStart ] = useState(-1)
-  const [ indicator, setIndicator ] = useState(null)
+  const [ trackingMouse, setTrackingMouse ] = useState(false);
+  const [ trackStart, setTrackStart ] = useState(-1);
+  const [ indicator, setIndicator ] = useState(null);
 
   const onMouseDown = (event) => {
-    const target = event.target
-    const indicatorElement = target.querySelector(".new-reservation-indicator")
+    const target = event.target;
+    const indicatorElement = target.querySelector(".new-reservation-indicator");
 
     const rect = target.getBoundingClientRect();
     const blockPosition = (event.clientX - rect.left);
 
-    indicatorElement.style.display = "block"
-    indicatorElement.style.left = `${blockPosition}px`
-    indicatorElement.style.width = `${0}px`
+    indicatorElement.style.display = "block";
+    indicatorElement.style.left = `${blockPosition}px`;
+    indicatorElement.style.width = `${0}px`;
 
-    setIndicator(indicatorElement)
-    setTrackStart(blockPosition)
-    setTrackingMouse(true)
+    setIndicator(indicatorElement);
+    setTrackStart(blockPosition);
+    setTrackingMouse(true);
   }
 
-  const onMouseUp = (event) => {
+  const onMouseUp = () => {
     // open modal here
-    setTrackStart(0)
-    setTrackingMouse(false)
-    setIndicator(null)
+    setTrackStart(0);
+    setTrackingMouse(false);
+    setIndicator(null);
   }
 
   const onMouseMove = (event) => {
     if (!trackingMouse) {
-      return
+      return;
     }
 
-    const target = event.target
+    const target = event.target;
 
     const rect = target.getBoundingClientRect();
     const blockSize = (event.clientX - trackStart - rect.left);
 
     if (blockSize > 0) {
-      indicator.style.width = `${blockSize}px`
+      indicator.style.width = `${blockSize}px`;
     } else {
-      indicator.style.left = `${trackStart - Math.abs(blockSize)}px`
-      indicator.style.width = `${Math.abs(blockSize)}px`
+      indicator.style.left = `${trackStart - Math.abs(blockSize)}px`;
+      indicator.style.width = `${Math.abs(blockSize)}px`;
     }
-
   }
 
   const stopEventPropagation = (event) => {
-    event.stopPropagation()
+    event.stopPropagation();
   }
 
   return (
@@ -138,6 +91,7 @@ const TimelineContainer = ({
             <div className="timelines col-start-2 row-span-full rounded-lg grid" style={{ gridTemplateRows: `30px repeat(${rooms.length}, 1fr)`} }>
               <div className="timelines__timestamps grid items-center justify-center top-0 bg-white z-20 border-b-2 sticky" style={{ gridTemplateColumns: "repeat(24, 1fr)"}}>
                 {
+                  /* Render timestamps */
                   Array.from({ length: 23 }).map((_, index) => (
                     <p
                       className="text-center bg-white translate-x-[-50%]"
@@ -149,46 +103,8 @@ const TimelineContainer = ({
                 }
               </div>
 
-
               { /* TIMELINE ITEM CONTAINER (this is the "actual" timeline */
-                rooms?.map((room) => (
-                  <div
-                    className="grid grid-flow-dense gap-x-2 gap-y-1 p-1 relative [&:not(:last-child)]:border-b border-b-gray-200 min-h-16" 
-                    style={{ gridTemplateRows: "1fr 1fr", gridTemplateColumns: `repeat(${24 * 4}, 1fr)`, width: `${24 * 4 * 30}px` }}
-                    onMouseDown={onMouseDown}
-                    onMouseUp={onMouseUp}
-                    onMouseMove={onMouseMove}
-                  >
-                    <div className="timelines__timestamp-lines absolute w-full h-full grid pointer-events-none" style={{ gridTemplateColumns: "repeat(24, 1fr" }}>
-                      {
-                        Array.from({ length: 24 }).map(() => (
-                          <div className="border-r border-gray-100 pointer-events-none"> </div>
-                        ))
-                      }
-                    </div>
-                    
-                    <TimelineItem timeStart="00:00" timeEnd="05:00" unavailable/>
-    
-                    {
-                      // Reservations inside the timeline
-                      room.reservations.map((reservation) => (
-                        <TimelineItem
-                          timeStart={reservation.startTime}
-                          timeEnd={reservation.endTime}
-                          label={reservation.purpose}
-                          user={reservation.user}
-                          fullHeight={reservation.groupsize === room.capacity}
-                        />
-                      ))
-                    }
-                    
-                    <TimelineItem timeStart="20:00" timeEnd="24:00" unavailable/>
-
-                    <div className="new-reservation-indicator absolute h-full bg-green-200 w-20 rounded-md pointer-events-none hidden">
-                      <p></p>
-                    </div>
-                  </div>
-                ))
+                rooms?.map((room) => <Timeline room={room} />)
               }
             </div>
           </>
@@ -198,71 +114,62 @@ const TimelineContainer = ({
   )
 }
 
-const Timeline = () => {
-  const { id } = useParams()
-  const { data, error, isLoading } = useSchoolQuery(id);  
+
+const TimelinePage = () => {
+  const { id } = useParams();
+  const { data, error, isLoading } = useSchoolQuery(id);
   const form = useForm({
     defaultValues: {
-      date: moment().format("MM/DD/YYYY") // ??
+      date: moment().format("MM/DD/YYYY") // Requires date in this format
     },
   });
 
-  const timelineContainerRef = useRef(null)
-  const timeIndicatorRef = useRef(null)
-  const timeIndicatorContainer = useRef(null)
+  setDefaultOptions({ locale: fi });
 
-  setDefaultOptions({ locale: fi })
+  const selectedBuilding = form.watch("building");
+  const selectedFloor = form.watch("floor");
+  const selectedDate = form.watch("date");
 
-  const selectedBuilding = form.watch("building")
-  const selectedFloor = form.watch("floor")
-  const selectedDate = form.watch("date")
+  const building = data.buildings.find((building) => building._id === selectedBuilding);
+  const floor = building?.floors.find((floor) => floor._id === selectedFloor);
 
-  const building = data.buildings.find((building) => building._id === selectedBuilding)
-  const floor = building?.floors.find((floor) => floor._id === selectedFloor)
-
-  let rooms = floor?.rooms
+  let rooms = floor?.rooms;
 
   rooms = floor?.rooms.map((room) => {
     const reservations = room.reservations.filter((reservation) => {
-      return moment(reservation.reservationDate).isSame(moment(selectedDate))
-    })
+      return moment(reservation.reservationDate).isSame(moment(selectedDate));
+    });
 
     return {
       ...room,
       reservations,
     }
-  }) 
+  });
 
   if (isLoading) {
     return <p>please wait</p>
   }
 
   return (
-    <div className="col-span-10 ml-5 mt-10">
-      <div className="flex flex-col gap-2">
-
-        <p className="text-4xl font-bold mb-3">{ data.name }</p>
-
-        {/* Filter */}
-        <TimelineFilterForm schoolData={data} form={form} />
-
-        <p className="text-3xl font-medium mt-6 mb-3">{ selectedDate && moment(selectedDate).format("LL") }</p>
-
-        <TimelineContainer
-          rooms={rooms}
-        />
-      </div>
-
-    </div>
-  )
-}
-
-const TimelinePage = () => {
-  return (
     <div className="grid grid-cols-10 gap-2">
-      <Timeline />
+      <div className="col-span-10 ml-5 mt-10">
+        <div className="flex flex-col gap-2">
+
+          <p className="text-4xl font-bold mb-3">{ data.name }</p>
+
+          {/* Filter */}
+          <TimelineFilterForm schoolData={data} form={form} />
+
+          <p className="text-3xl font-medium mt-6 mb-3">{ selectedDate && moment(selectedDate).format("LL") }</p>
+
+          <TimelineContainer
+            rooms={rooms}
+          />
+        </div>
+
+      </div>
     </div>
-  )
+  );
 }
 
-export default TimelinePage
+export default TimelinePage;
